add TPR, cleanup support to walkthrough e2e script

+++ contrib/hack/test_walkthrough.sh
@@ -25,43 +25,79 @@ while [[ $# -gt 0 ]]; do
   case "${1}" in
     --registry)   REGISTRY="${2:-}"; shift ;;
     --version)    VERSION="${2:-}"; shift ;;
-
+    --with-tpr)   WITH_TPR=1 ;;
+    --cleanup)    CLEANUP=1 ;;
     *) error_exit "Unrecognized command line parameter: $1" ;;
   esac
   shift
 done
 
+BROKER_RELEASE="ups-broker"
+CATALOG_RELEASE="catalog"
+
+K8S_KUBECONFIG="${KUBECONFIG:-~/.kube/config}"
+SC_KUBECONFIG="/tmp/sc-kubeconfig"
+
 VERSION="${VERSION:-"$(git describe --tags --always --abbrev=7 --dirty)"}" \
   || error_exit 'Cannot determine Git commit SHA'
 
+REGISTRY="${REGISTRY:-}"
+
+function cleanup() {
+  echo 'Cleaning up resources...'
+  export KUBECONFIG="${K8S_KUBECONFIG}"
+  helm delete --purge "${BROKER_RELEASE}" || true
+  helm delete --purge "${CATALOG_RELEASE}" || true
+  rm -f "${SC_KUBECONFIG}"
+  kubectl delete secret -n test-ns my-secret > /dev/null 2>&1 || true
+  kubectl delete namespace test-ns || true
+
+  wait_for_expected_output -x -e 'test-ns' -n 10 \
+    kubectl get namespaces
+
+  # TODO: Hack in order to delete TPRs. Will need to be removed when TPRs can be deleted
+  # by the catalog API server.
+  if [[ -n "${WITH_TPR:-}" ]]; then
+    kubectl delete thirdpartyresources binding.servicecatalog.k8s.io
+    kubectl delete thirdpartyresources instance.servicecatalog.k8s.io
+    kubectl delete thirdpartyresources broker.servicecatalog.k8s.io
+    kubectl delete thirdpartyresources service-class.servicecatalog.k8s.io
+  fi
+
+  echo 'Cleanup done.'
+}
+
 # Deploying to cluster
 
-export KUBECONFIG="${K8S_KUBECONFIG}"
+if [[ -n "${CLEANUP:-}" ]]; then
+  trap cleanup EXIT
+fi
+
 kubectl create namespace test-ns
 
 echo 'Deploying user-provided-service broker...'
 
-VALUES="version=${VERSION}"
-if [[ -n "${REGISTRY:-}" ]]; then
-  VALUES+=",registry=${REGISTRY}"
-fi
-
 retry -n 10 \
     helm install "${ROOT}/charts/ups-broker" \
-    --name "ups-broker" \
+    --name "${BROKER_RELEASE}" \
     --namespace "ups-broker" \
-    --set "${VALUES}" \
   || error_exit 'Error deploy ups broker to cluster.'
 
 echo 'Deploying service catalog...'
 
-VALUES+=',debug=true'
+VALUES='debug=true'
 VALUES+=',insecure=true'
+VALUES+=",controllerManager.image=${REGISTRY}controller-manager:${VERSION}"
+VALUES+=",apiserver.image=${REGISTRY}apiserver:${VERSION}"
 VALUES+=',apiserver.service.type=LoadBalancer'
+if [[ -n "${WITH_TPR:-}" ]]; then
+  VALUES+=',apiserver.storage.type=tpr'
+  VALUES+=',apiserver.storage.tpr.globalNamespace=test-ns'
+fi
 
 retry -n 10 \
     helm install "${ROOT}/charts/catalog" \
-    --name "catalog" \
+    --name "${CATALOG_RELEASE}" \
     --namespace "catalog" \
     --set "${VALUES}" \
   || error_exit 'Error deploying service catalog to cluster.'
@@ -74,16 +110,16 @@ wait_for_expected_output -x -e 'ContainerCreating' -n 10 \
     kubectl get pods --namespace ups-broker \
   || error_exit 'User provided service broker pod took an unexpected amount of time to come up.'
 
-[[ "$(kubectl get pods --namespace ups-broker | grep ups-broker | awk '{print $3}')" == 'Running' ]] \
+[[ "$(kubectl get pods --namespace ups-broker | grep ups-broker-ups-broker | awk '{print $3}')" == 'Running' ]] \
   || error_exit 'User provided service broker pod did not come up successfully.'
 
 wait_for_expected_output -x -e 'ContainerCreating' -n 10 \
     kubectl get pods --namespace catalog \
   || error_exit 'Service catalog pods did not come up successfully.'
 
-[[ "$(kubectl get pods --namespace catalog | grep apiserver | awk '{print $3}')" == 'Running' ]] \
+[[ "$(kubectl get pods --namespace catalog | grep catalog-catalog-apiserver | awk '{print $3}')" == 'Running' ]] \
   || error_exit 'API server pod did not come up successfully.'
-[[ "$(kubectl get pods --namespace catalog | grep controller | awk '{print $3}')" == 'Running' ]] \
+[[ "$(kubectl get pods --namespace catalog | grep catalog-catalog-controller | awk '{print $3}')" == 'Running' ]] \
   || error_exit 'Controller pod did not come up successfully.'
 
 echo 'Waiting on external IP for service catalog API Server...'
@@ -108,7 +144,8 @@ kubectl config set-cluster service-catalog-cluster --server="http://${API_SERVER
 kubectl config set-context service-catalog-ctx --cluster=service-catalog-cluster --user=service-catalog-creds
 kubectl config use-context service-catalog-ctx
 
-[[ "$(kubectl get brokers,serviceclasses,instances,bindings 2>&1)" == 'No resources found.' ]] \
+retry -n 10 \
+  kubectl get brokers,serviceclasses,instances,bindings \
   || error_exit 'Issue listing resources from service catalog API server.'
 
 # Create the broker
@@ -159,38 +196,43 @@ wait_for_expected_output -e 'InjectedBindResult' -n 10 \
 [[ "$(KUBECONFIG="${K8S_KUBECONFIG}" kubectl get secrets -n test-ns)" == *my-secret* ]] \
   || error_exit '"my-secret" not present when listing secrets.'
 
-# Unbind from the instance
 
-echo 'Unbinding from instance...'
+# TODO: Cannot currently test TPR deletion; only delete if using an etcd-backed
+# API server
+if [[ -z "${WITH_TPR:-}" ]]; then
+  # Unbind from the instance
 
-kubectl delete -n test-ns bindings ups-binding \
-  || error_exit 'Error when deleting ups-binding.'
+  echo 'Unbinding from instance...'
 
-export KUBECONFIG="${K8S_KUBECONFIG}"
-wait_for_expected_output -x -e "my-secret" -n 10 \
-    kubectl get secrets -n test-ns \
-  || error_exit '"my-secret" not removed upon deleting ups-binding.'
-export KUBECONFIG="${SC_KUBECONFIG}"
+  kubectl delete -n test-ns bindings ups-binding \
+    || error_exit 'Error when deleting ups-binding.'
 
-# Deprovision the instance
+  export KUBECONFIG="${K8S_KUBECONFIG}"
+  wait_for_expected_output -x -e "my-secret" -n 10 \
+      kubectl get secrets -n test-ns \
+    || error_exit '"my-secret" not removed upon deleting ups-binding.'
+  export KUBECONFIG="${SC_KUBECONFIG}"
 
-echo 'Deprovisioning instance...'
+  # Deprovision the instance
 
-kubectl delete -n test-ns instances ups-instance \
-  || error_exit 'Error when deleting ups-instance.'
+  echo 'Deprovisioning instance...'
 
-# Delete the broker
+  kubectl delete -n test-ns instances ups-instance \
+    || error_exit 'Error when deleting ups-instance.'
 
-echo 'Deleting broker...'
+  # Delete the broker
 
-kubectl delete brokers ups-broker \
-  || error_exit 'Error when deleting ups-broker.'
+  echo 'Deleting broker...'
 
-wait_for_expected_output -x -e 'user-provided-service' -n 10 \
-    kubectl get serviceclasses \
-  || error_exit 'Service classes not successfully removed upon deleting ups-broker.'
+  kubectl delete brokers ups-broker \
+    || error_exit 'Error when deleting ups-broker.'
 
-[[ "$(kubectl get serviceclasses 2>&1)" == "No resources found." ]] \
-  || error_exit 'Service classes not successfully removed upon deleting ups-broker.'
+  wait_for_expected_output -x -e 'user-provided-service' -n 10 \
+      kubectl get serviceclasses \
+    || error_exit 'Service classes not successfully removed upon deleting ups-broker.'
+
+  [[ "$(kubectl get serviceclasses 2>&1)" == "No resources found." ]] \
+    || error_exit 'Service classes not successfully removed upon deleting ups-broker.'
+fi
 
 echo 'Walkthrough completed successfully.'
